update for using arrow functions
